Refactor: Update Create task endpoint to support multiple tags
This commit modifies the Create task endpoint in the TasksController to support the addition of multiple tags to a task. The previous implementation only allowed for a single tag to be added. Now, an array of tag names can be provided in the request body, and the endpoint will create the task with all the specified tags.

The code changes include:
- Adding a new variable `tagNames` to store the array of tag names
- Updating the `dto` object to use the `tagNames` array instead of a single tag name
- Modifying the assertion in the test case to check that all the specified tags are present in the response body